Move file in leaderboard dedicated folder
